+++ src/components/InfographicEditor/Editors/PieEditor.js
@@ -219,16 +219,29 @@ class PieEditor extends React.Component
 
     _GetSizeContents()
     {
-        let size = this.props.cSettings.size;
+        let size = this.props.cSettings.size, 
+            diff = size.chartRadius - size.innerRadius,
+            sliderValue = (diff <= 0) ? 0 : diff;
         let innerRadiusContent = (this.props.type === 'pie-editor') ? false : 
-            ( <LabeledNumericTextField 
-                label='Thickness:'
-                index='donut-radius'
-                initialValue={parseFloat(size.chartRadius) - parseFloat(size.innerRadius)}
-                rows={1}
-                cols={5}
-                onChange={ (d, i) => { this._SetDonutRadius(d); }}
-            />);
+            ( 
+                <div style={{
+                    display: 'grid', gridTemplateColumns: 'auto auto', 
+                    gridColumnGap: '20px'
+                }}>
+                    <label>Thickness:</label>
+                    <input 
+                        type='range'
+                        style={{width: 'auto'}}
+                        min={10} 
+                        max={size.chartRadius}
+                        step={1}
+                        value={sliderValue}
+                        disabled={size.chartRadius <= 20}
+                        onChange={event => { 
+                            this._SetDonutRadius(event.target.value)}}
+                    ></input>
+                </div>
+            );
         return [
             <div className='center'>
                 <LabeledNumericTextField 
@@ -253,31 +266,6 @@ class PieEditor extends React.Component
         let statistic = (isDisabled) ? this._defaultSettings.statistic : this.props.dSettings.statistic;
         return [
             <div className='center'>
-                <div>
-                    <h5>Position Settings:</h5>
-                    <LabeledTextField 
-                        label='X:'
-                        index='label-x'
-                        initialValue={statistic.position.x}
-                        rows={1}
-                        cols={5}
-                        onChange={(d, i) => { 
-                            this._UpdateDecoratorSettings('statistic', 'position', 'x', d);
-                        }} 
-                        isDisabled={isDisabled}
-                    />
-                    <LabeledTextField 
-                        label='Y:'
-                        index='label-y'
-                        initialValue={statistic.position.y}
-                        rows={1}
-                        cols={5}
-                        onChange={(d, i) => { 
-                            this._UpdateDecoratorSettings('statistic', 'position', 'y', d);
-                        }} 
-                        isDisabled={isDisabled}
-                    />
-                </div>
                 <div>
                     <h5>Font Settings:</h5>
                     <FontSelector